+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "kasuki"
-version = "2.8.6"
+version = "2.8.7"
 edition = "2021"
 authors = ["Valgul <contact@valgul.moe>"]
 description = "A discord bot written in rust that get info from anilist API"
@@ -13,7 +13,7 @@ license-file = "LICENSE"
 [dependencies]
 regex = "1.10.4"
 reqwest = { version = "0.12.4", features = ["json", "multipart", "blocking"] }
-serde = "1.0.199"
+serde = "1.0.200"
 serde_json = "1.0.116"
 serenity = { version = "0.12.1", default-features = false, features = ["client", "gateway", "rustls_backend", "model", "utils", "full"] }
 tokio = { version = "1.37.0", features = ["full"] }
@@ -23,7 +23,7 @@ chrono = "0.4.38"
 uuid = { version = "1.8.0", features = ["v4"] }
 dotenvy = "0.15.7"
 image = "0.25.1"
-base64 = "0.22.0"
+base64 = "0.22.1"
 tracing = { version = "0.1.40" }
 once_cell = "1.19.0"
 tracing-subscriber = { version = "0.3.18", features = ["default", "env-filter"] }
