+++ src/command_register/registration_function/register_subcommand.rs
@@ -6,9 +6,20 @@ use serenity::all::{CommandType, CreateCommand, Http, Permissions};
 use tracing::{error, trace};
 
 use crate::command_register::command_struct::subcommand::SubCommand;
-use crate::command_register::registration_function::common::get_subcommand_option;
+use crate::command_register::registration_function::common::{
+    get_permission, get_subcommand_option,
+};
 use crate::error_management::error_enum::{AppError, ErrorResponseType, ErrorType};
 
+/// This asynchronous function creates subcommands in Discord by reading from a JSON file and sending them to the Discord API.
+///
+/// It first calls the `get_subcommands` function to read the subcommands from the JSON file located at "./json/subcommand".
+/// If an error occurs during this process, it logs the error and returns early.
+/// If the subcommands are successfully read, it iterates over each subcommand and calls the `create_command` function to send the subcommand to the Discord API.
+///
+/// # Arguments
+///
+/// * `http` - An `Arc<Http>` instance used to send the subcommands to the Discord API.
 pub async fn creates_subcommands(http: &Arc<Http>) {
     let commands = match get_subcommands("./json/subcommand") {
         Err(e) => {
@@ -23,6 +34,22 @@ pub async fn creates_subcommands(http: &Arc<Http>) {
     }
 }
 
+/// This function reads subcommands from a JSON file located at the given path and returns them as a vector of `SubCommand` structs.
+///
+/// It first reads the directory at the given path and maps any errors to an `AppError`.
+/// It then iterates over each entry in the directory.
+/// If an entry is a file with a ".json" extension, it opens the file and reads it into a `SubCommand` struct.
+/// If an error occurs during this process, it maps the error to an `AppError`.
+/// If the subcommand is successfully read, it is pushed to the `subcommands` vector.
+/// If no subcommands are found in the directory, it logs a trace message.
+///
+/// # Arguments
+///
+/// * `path` - A string slice representing the path to the directory containing the JSON files.
+///
+/// # Returns
+///
+/// A `Result` containing either a vector of `SubCommand` structs if the subcommands are successfully read, or an `AppError` if an error occurs.
 fn get_subcommands(path: &str) -> Result<Vec<SubCommand>, AppError> {
     let mut subcommands = Vec::new();
     let paths = fs::read_dir(path).map_err(|e| AppError {
@@ -63,6 +90,20 @@ fn get_subcommands(path: &str) -> Result<Vec<SubCommand>, AppError> {
     Ok(subcommands)
 }
 
+/// This asynchronous function creates a global subcommand in Discord using the provided `SubCommand` struct and `Http` instance.
+///
+/// It first creates a `CreateCommand` instance using the name, NSFW status, command type, DM permission status, and description from the `SubCommand` struct.
+///
+/// It then calls the `get_permission` function to set the default member permissions of the `CreateCommand` based on the permissions in the `SubCommand` struct.
+///
+/// If the `SubCommand` struct contains a command, it calls the `get_subcommand_option` function to convert it into command options and sets them on the `CreateCommand`.
+///
+/// Finally, it sends the `CreateCommand` to the Discord API to create the global subcommand. If an error occurs during this process, it logs the error.
+///
+/// # Arguments
+///
+/// * `command` - A reference to a `SubCommand` struct containing the details of the subcommand to be created.
+/// * `http` - An `Arc<Http>` instance used to send the subcommand to the Discord API.
 async fn create_command(command: &SubCommand, http: &Arc<Http>) {
     let mut command_build = CreateCommand::new(&command.name)
         .nsfw(command.nsfw)
@@ -70,18 +111,8 @@ async fn create_command(command: &SubCommand, http: &Arc<Http>) {
         .dm_permission(command.dm_command)
         .description(&command.desc);
 
-    command_build = match &command.permissions {
-        Some(permissions) => {
-            let mut perm_bit: u64 = 0;
-            for perm in permissions {
-                let permission: Permissions = perm.permission.into();
-                perm_bit |= permission.bits()
-            }
-            let permission = Permissions::from_bits(perm_bit).unwrap();
-            command_build.default_member_permissions(permission)
-        }
-        None => command_build,
-    };
+    command_build = get_permission(&command.permissions, command_build);
+
     command_build = match &command.command {
         Some(command) => {
             let options = get_subcommand_option(command);
