+++ src/command_register/registration_function/register_subcommand_group.rs
@@ -7,10 +7,19 @@ use tracing::{error, trace};
 
 use crate::command_register::command_struct::subcommand_group::SubCommandGroup;
 use crate::command_register::registration_function::common::{
-    get_subcommand_group_option, get_subcommand_option,
+    get_permission, get_subcommand_group_option, get_subcommand_option,
 };
 use crate::error_management::error_enum::{AppError, ErrorResponseType, ErrorType};
 
+/// This asynchronous function creates subcommand groups in Discord by reading from a JSON file and sending them to the Discord API.
+///
+/// It first calls the `get_subcommands_group` function to read the subcommand groups from the JSON file located at "./json/subcommand_group".
+/// If an error occurs during this process, it logs the error and returns early.
+/// If the subcommand groups are successfully read, it iterates over each subcommand group and calls the `create_command` function to send the subcommand group to the Discord API.
+///
+/// # Arguments
+///
+/// * `http` - An `Arc<Http>` instance used to send the subcommand groups to the Discord API.
 pub async fn creates_subcommands_group(http: &Arc<Http>) {
     let commands = match get_subcommands_group("./json/subcommand_group") {
         Err(e) => {
@@ -25,6 +34,22 @@ pub async fn creates_subcommands_group(http: &Arc<Http>) {
     }
 }
 
+/// This function reads subcommand groups from a JSON file located at the given path and returns them as a vector of `SubCommandGroup` structs.
+///
+/// It first reads the directory at the given path and maps any errors to an `AppError`.
+/// It then iterates over each entry in the directory.
+/// If an entry is a file with a ".json" extension, it opens the file and reads it into a `SubCommandGroup` struct.
+/// If an error occurs during this process, it maps the error to an `AppError`.
+/// If the subcommand group is successfully read, it is pushed to the `subcommands_group` vector.
+/// If no subcommand groups are found in the directory, it logs a trace message.
+///
+/// # Arguments
+///
+/// * `path` - A string slice representing the path to the directory containing the JSON files.
+///
+/// # Returns
+///
+/// A `Result` containing either a vector of `SubCommandGroup` structs if the subcommand groups are successfully read, or an `AppError` if an error occurs.
 fn get_subcommands_group(path: &str) -> Result<Vec<SubCommandGroup>, AppError> {
     let mut subcommands_group = Vec::new();
     let paths = fs::read_dir(path).map_err(|e| AppError {
@@ -66,6 +91,22 @@ fn get_subcommands_group(path: &str) -> Result<Vec<SubCommandGroup>, AppError> {
     Ok(subcommands_group)
 }
 
+/// This asynchronous function creates a global subcommand group in Discord using the provided `SubCommandGroup` struct and `Http` instance.
+///
+/// It first creates a `CreateCommand` instance using the name, NSFW status, command type, DM permission status, and description from the `SubCommandGroup` struct.
+///
+/// It then calls the `get_permission` function to set the default member permissions of the `CreateCommand` based on the permissions in the `SubCommandGroup` struct.
+///
+/// If the `SubCommandGroup` struct contains a command, it calls the `get_subcommand_option` function to convert it into command options and sets them on the `CreateCommand`.
+///
+/// If the `SubCommandGroup` struct contains subcommands, it calls the `get_subcommand_group_option` function to convert them into command options and sets them on the `CreateCommand`.
+///
+/// Finally, it sends the `CreateCommand` to the Discord API to create the global subcommand group. If an error occurs during this process, it logs the error.
+///
+/// # Arguments
+///
+/// * `command` - A reference to a `SubCommandGroup` struct containing the details of the subcommand group to be created.
+/// * `http` - An `Arc<Http>` instance used to send the subcommand group to the Discord API.
 async fn create_command(command: &SubCommandGroup, http: &Arc<Http>) {
     let mut command_build = CreateCommand::new(&command.name)
         .nsfw(command.nsfw)
@@ -73,18 +114,7 @@ async fn create_command(command: &SubCommandGroup, http: &Arc<Http>) {
         .dm_permission(command.dm_command)
         .description(&command.desc);
 
-    command_build = match &command.permissions {
-        Some(permissions) => {
-            let mut perm_bit: u64 = 0;
-            for perm in permissions {
-                let permission: Permissions = perm.permission.into();
-                perm_bit |= permission.bits()
-            }
-            let permission = Permissions::from_bits(perm_bit).unwrap();
-            command_build.default_member_permissions(permission)
-        }
-        None => command_build,
-    };
+    command_build = get_permission(&command.permissions, command_build);
 
     command_build = match &command.command {
         Some(command) => {
